# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My Digimon Project CRUD",
    "description": "My Digimon Project Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3100",
  "basePath": "/",
  "tags": [
    {
      "name": "Cards",
      "description": "API for consult Digimon Cards Game"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/cards": {
      "get": {
        "tags": ["Cards"],
        "summary": "Get all cards in Mongo",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "limit of the results"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Card"
            }
          }
        }
      }
    },
    "/api/card/{cardnumber}": {
      "get": {
        "tags": ["Cards"],
        "summary": "Get a card of Mongo",
        "parameters": [
          {
            "name": "cardnumber",
            "in": "path",
            "description": "number of card"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Card"
            }
          }
        }
      }
    },
    "/api/card": {
      "post": {
        "tags": ["Card"],
        "description": "Create new card in DB",
        "parameters": [
          {
            "name": "card",
            "in": "body",
            "description": "Card that we want to create",
            "schema": {
              "$ref": "#/definitions/Card_Without_ID"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New card is created",
            "schema": {
              "$ref": "#/definitions/Card"
            }
          }
        }
      }
    },
    "/api/card/{id}": {
      "put": {
        "tags": ["Card"],
        "description": "Update the quantity of one card",
        "parameters": [
          {
            "name": "card",
            "in": "body",
            "description": "Card that we want to create",
            "schema": {
              "$ref": "#/definitions/Quantity"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Id of the card"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New card is created",
            "schema": {
              "$ref": "#/definitions/Card"
            }
          }
        }
      }
    },
    "/api/card/{cardId}": {
        "delete": {
          "tags": ["Card"],
          "description": "Update the quantity of one card",
          "parameters": [
            {
              "name": "cardId",
              "in": "path",
              "description": "Id of the card"
            }
          ],
          "produces": ["application/json"],
          "responses": {
            "204": {
              "description": "The card is deleted"
            }
          }
        }
      }
  },
  "definitions": {
    "Card": {
      "required": ["name", "_id"],
      "properties": {
        "id": {
          "type": "string",
          "uniqueItems": true
        },
        "name": {
          "type": "string"
        },
        "color": {
          "type": "string"
        },
        "stage": {
          "type": "string"
        },
        "digi_type": {
          "type": "string"
        },
        "attribute": {
          "type": "string"
        },
        "level": {
          "type": "number"
        },
        "play_cost": {
          "type": "number"
        },
        "evolution_cost": {
          "type": "number"
        },
        "cardrarety": {
          "type": "string"
        },
        "artist": {
          "type": "string"
        },
        "dp": {
          "type": "number"
        },
        "cardnumber": {
          "type": "string"
        },
        "maineffect": {
          "type": "string"
        },
        "soureeffect": {
          "type": "string"
        },
        "set_name": {
          "type": "string"
        },
        "image_url": {
          "type": "string"
        },
        "quantity": {
          "type": "number"
        },
        "card_sets": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "Card_Without_ID": {
      "required": ["name"],
      "properties": {
        "name": {
          "type": "string"
        },
        "color": {
          "type": "string"
        },
        "stage": {
          "type": "string"
        },
        "digi_type": {
          "type": "string"
        },
        "attribute": {
          "type": "string"
        },
        "level": {
          "type": "number"
        },
        "play_cost": {
          "type": "number"
        },
        "evolution_cost": {
          "type": "number"
        },
        "cardrarety": {
          "type": "string"
        },
        "artist": {
          "type": "string"
        },
        "dp": {
          "type": "number"
        },
        "cardnumber": {
          "type": "string"
        },
        "maineffect": {
          "type": "string"
        },
        "soureeffect": {
          "type": "string"
        },
        "set_name": {
          "type": "string"
        },
        "image_url": {
          "type": "string"
        },
        "quantity": {
          "type": "number"
        },
        "card_sets": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "Quantity": {
      "required": ["name"],
      "properties": {
        "name": {
          "type": "string"
        },
        "quantity": {
          "type": "number"
        }
      }
    }
  }
}